See changelog 2024-01-16

+++ mrerr/error_kind.go
@@ -8,7 +8,7 @@ const (
 )
 
 type (
-	ErrorKind int
+	ErrorKind int8
 )
 
 func (k ErrorKind) String() string {
